+++ R/tokenize.R
@@ -14,14 +14,16 @@
 #' @param columns A list of tibble results that define the
 #'  encoding. This is `NULL` until the step is trained by
 #'  [recipes::prep.recipe()].
+#' @param training_options A list of options passed to the tokenizer when it is
+#'  being trained. Only applicable for engine == "tokenizers.bpe".
 #' @param options A list of options passed to the tokenizer.
-#' @param token Unit for tokenizing. See details for options. Defaults to 
+#' @param token Unit for tokenizing. See details for options. Defaults to
 #' "words".
-#' @param engine Package that will be used for tokenization. See details for 
+#' @param engine Package that will be used for tokenization. See details for
 #' options. Defaults to "tokenizers".
 #' @param custom_token User supplied tokenizer. Use of this argument
 #'  will overwrite the token and engine arguments. Must take a character vector
-#'  as input and output a list of character vectors. 
+#'  as input and output a list of character vectors.
 #' @param skip A logical. Should the step be skipped when the
 #'  recipe is baked by [recipes::bake.recipe()]? While all
 #'  operations are baked when [recipes::prep.recipe()] is run, some
@@ -38,41 +40,41 @@
 #' library(recipes)
 #' library(modeldata)
 #' data(okc_text)
-#' 
-#' okc_rec <- recipe(~ ., data = okc_text) %>%
-#'   step_tokenize(essay0) 
-#'   
+#'
+#' okc_rec <- recipe(~., data = okc_text) %>%
+#'   step_tokenize(essay0)
+#'
 #' okc_obj <- okc_rec %>%
 #'   prep()
-#' 
-#' juice(okc_obj, essay0) %>%
+#'
+#' bake(okc_obj, new_data = NULL, essay0) %>%
 #'   slice(1:2)
-#' 
-#' juice(okc_obj) %>%
+#'
+#' bake(okc_obj, new_data = NULL) %>%
 #'   slice(2) %>%
 #'   pull(essay0)
-#'   
+#'
 #' tidy(okc_rec, number = 1)
 #' tidy(okc_obj, number = 1)
-#' 
-#' okc_obj_chars <- recipe(~ ., data = okc_text) %>%
+#'
+#' okc_obj_chars <- recipe(~., data = okc_text) %>%
 #'   step_tokenize(essay0, token = "characters") %>%
 #'   prep()
-#' 
-#' juice(okc_obj_chars) %>%
+#'
+#' bake(okc_obj, new_data = NULL) %>%
 #'   slice(2) %>%
 #'   pull(essay0)
 #' @export
-#' @details 
+#' @details
 #' Tokenization is the act of splitting a character string into smaller parts
-#' to be further analysed. This step uses the `tokenizers` package which 
+#' to be further analysed. This step uses the `tokenizers` package which
 #' includes heuristics to split the text into paragraphs tokens, word tokens
 #' amoug others. `textrecipes` keeps the tokens in a [tokenlist] and other
 #' steps will do their tasks on those [tokenlist]s before transforming them
 #' back to numeric.
-#' 
+#'
 #' The choice of `engine` determines the possible choices of `token`.
-#' 
+#'
 #' If `engine = "tokenizers"`:
 #' * "words" (default)
 #' * "characters"
@@ -87,11 +89,11 @@
 #' * "ptb" (Penn Treebank)
 #' * "skip_ngrams"
 #' * "word_stems"
-#' 
+#'
 #' if `engine = "spacyr"`
 #' * "words"
-#' 
-#' Working will `textrecipes` will almost always start by calling 
+#'
+#' Working will `textrecipes` will almost always start by calling
 #' `step_tokenize` followed by modifying and filtering steps. This is not always
 #' the case as you sometimes want to do apply pre-tokenization steps, this can
 #' be done with [recipes::step_mutate()].
@@ -104,13 +106,13 @@ step_tokenize <-
            role = NA,
            trained = FALSE,
            columns = NULL,
+           training_options = list(),
            options = list(),
            token = "words",
            engine = "tokenizers",
            custom_token = NULL,
            skip = FALSE,
-           id = rand_id("tokenize")
-  ) {
+           id = rand_id("tokenize")) {
     add_step(
       recipe,
       step_tokenize_new(
@@ -118,6 +120,7 @@ step_tokenize <-
         role = role,
         trained = trained,
         columns = columns,
+        training_options = training_options,
         options = options,
         token = token,
         engine = engine,
@@ -129,14 +132,15 @@ step_tokenize <-
   }
 
 step_tokenize_new <-
-  function(terms, role, trained, columns, options, token, engine, custom_token,
-           skip, id) {
+  function(terms, role, trained, columns, training_options, options, token,
+           engine, custom_token, skip, id) {
     step(
       subclass = "tokenize",
       terms = terms,
       role = role,
       trained = trained,
       columns = columns,
+      training_options = training_options,
       options = options,
       token = token,
       engine = engine,
@@ -154,15 +158,23 @@ prep.step_tokenize <- function(x, training, info = NULL, ...) {
 
   check_type(training[, col_names], quant = FALSE)
 
+  tokenizers <- list()
+
+  for (i in seq_along(col_names)) {
+    tokenizers[[i]] <- x$custom_token %||%
+      tokenizer_switch(x$token, x, training[, col_names[[i]], drop = TRUE])
+  }
+
   step_tokenize_new(
     terms = x$terms,
     role = x$role,
     trained = TRUE,
     columns = col_names,
+    training_options = x$training_options,
     options = x$options,
     token = x$token,
     engine = x$engine,
-    custom_token = x$custom_token,
+    custom_token = tokenizers,
     skip = x$skip,
     id = x$id
   )
@@ -173,14 +185,13 @@ bake.step_tokenize <- function(object, new_data, ...) {
   col_names <- object$columns
   # for backward compat
 
-  tokenizer <- object$custom_token %||%
-                           tokenizer_switch(object$token, object$engine)
-
   for (i in seq_along(col_names)) {
-    new_data[, col_names[i]] <- tokenizer_fun(new_data[, col_names[i]],
-                                              col_names[i],
-                                              options = object$options,
-                                              token = tokenizer)
+    new_data[, col_names[i]] <- tokenizer_fun(
+      data = new_data[, col_names[i]],
+      name = col_names[i],
+      options = object$options,
+      token = object$custom_token[[i]]
+    )
   }
   as_tibble(new_data)
 }
@@ -198,12 +209,16 @@ print.step_tokenize <-
 #' @export
 tidy.step_tokenize <- function(x, ...) {
   if (is_trained(x)) {
-    res <- tibble(terms = x$terms,
-                  value = x$token)
+    res <- tibble(
+      terms = x$terms,
+      value = x$token
+    )
   } else {
     term_names <- sel2char(x$terms)
-    res <- tibble(terms = term_names,
-                  value = na_chr)
+    res <- tibble(
+      terms = term_names,
+      value = na_chr
+    )
   }
   res$id <- x$id
   res
@@ -212,20 +227,21 @@ tidy.step_tokenize <- function(x, ...) {
 ## Implementation
 tokenizer_fun <- function(data, name, options, token, ...) {
   check_type(data[, name], quant = FALSE)
-  
+
   data <- factor_to_text(data, name)
-  
+
   token_expr <- expr(
     token(
       x = data[, 1, drop = TRUE]
     )
   )
-  
-  if (length(options) > 0)
+
+  if (length(options) > 0) {
     token_expr <- mod_call_args(token_expr, args = options)
+  }
 
   token_list <- eval(token_expr)
-  
+
   if (is_tokenlist(token_list)) {
     out <- tibble::tibble(token_list)
   } else {
@@ -235,48 +251,129 @@ tokenizer_fun <- function(data, name, options, token, ...) {
   out
 }
 
-tokenizer_switch <- function(name, engine) {
-  if (engine == "tokenizers") {
+tokenizer_switch <- function(name, object, data) {
+  if (object$engine == "tokenizers") {
     possible_tokenizers <-
-      c("characters", "character_shingle", "lines", "ngrams",
+      c(
+        "characters", "character_shingle", "lines", "ngrams",
         "paragraphs", "ptb", "regex", "sentences", "skip_ngrams",
-        "tweets", "words", "word_stems")
-    
-    if (!(name %in% possible_tokenizers))
+        "tweets", "words", "word_stems"
+      )
+
+    if (!(name %in% possible_tokenizers)) {
       rlang::abort(paste0("token should be one of the supported ",
-                          "'", possible_tokenizers, "'", collapse = ", "))
-    
+        "'", possible_tokenizers, "'",
+        collapse = ", "
+      ))
+    }
+
     res <- switch(name,
-           characters = tokenizers::tokenize_characters,
-           character_shingle = tokenizers::tokenize_character_shingles,
-           lines = tokenizers::tokenize_lines,
-           ngrams = tokenizers::tokenize_ngrams,
-           paragraphs = tokenizers::tokenize_paragraphs,
-           ptb = tokenizers::tokenize_ptb,
-           regex = tokenizers::tokenize_regex,
-           sentences = tokenizers::tokenize_sentences,
-           skip_ngrams = tokenizers::tokenize_skip_ngrams,
-           tweets = tokenizers::tokenize_tweets,
-           words = tokenizers::tokenize_words,
-           word_stems = tokenizers::tokenize_word_stems
+      characters = tokenizers::tokenize_characters,
+      character_shingle = tokenizers::tokenize_character_shingles,
+      lines = tokenizers::tokenize_lines,
+      ngrams = tokenizers::tokenize_ngrams,
+      paragraphs = tokenizers::tokenize_paragraphs,
+      ptb = tokenizers::tokenize_ptb,
+      regex = tokenizers::tokenize_regex,
+      sentences = tokenizers::tokenize_sentences,
+      skip_ngrams = tokenizers::tokenize_skip_ngrams,
+      tweets = tokenizers::tokenize_tweets,
+      words = tokenizers::tokenize_words,
+      word_stems = tokenizers::tokenize_word_stems
     )
     return(res)
   }
-  
-  if (engine == "spacyr") {
-    recipes::recipes_pkg_check("spacyr")
-    
+
+  if (object$engine == "spacyr") {
+    recipes::recipes_pkg_check(required_pkgs.step_tokenize(object))
+
     possible_tokenizers <- c("words")
-    
-    if (!(name %in% possible_tokenizers))
-      rlang::abort(paste0("token should be one of the supported ",
-                          "'", possible_tokenizers, "'", collapse = ", "))
-    
+
+    if (!(name %in% possible_tokenizers)) {
+      rlang::abort(paste0(
+        "token should be one of the supported ",
+        "'", 
+        possible_tokenizers, 
+        "'",
+        collapse = ", "
+      ))
+    }
+
     res <- switch(name,
-                  words = spacyr_tokenizer_words
+      words = spacyr_tokenizer_words
     )
     return(res)
   }
-  
+
+  if (object$engine == "tokenizers.bpe") {
+    recipes::recipes_pkg_check(required_pkgs.step_tokenize(object))
+
+    possible_tokenizers <- c("words")
+
+    if (!(name %in% possible_tokenizers)) {
+      rlang::abort(paste0(
+        "token should be one of the supported ",
+        "'",
+        possible_tokenizers,
+        "'",
+        collapse = ", "
+      ))
+    }
+
+    res <- switch(name,
+      words = tokenizers_bpe_words(data, object$training_options)
+    )
+    return(res)
+  }
+
+  if (object$engine == "udpipe") {
+    recipes::recipes_pkg_check(required_pkgs.step_tokenize(object))
+
+    possible_tokenizers <- c("words")
+
+    if (!(name %in% possible_tokenizers)) {
+      rlang::abort(paste0(
+        "token should be one of the supported ",
+        "'",
+        possible_tokenizers,
+        "'",
+        collapse = ", "
+      ))
+    }
+
+    res <- switch(name,
+      words = udpipe_words(object$training_options$model)
+    )
+    return(res)
+  }
+
   rlang::abort("`engine` argument is not valid.")
+}
+
+#' @rdname required_pkgs.step
+#' @export
+required_pkgs.step_tokenize <- function(x, ...) {
+  if (x$engine == "spacyr") {
+    c("spacyr", "textrecipes")
+  } else if (x$engine == "tokenizers.bpe") {
+    c("tokenizers.bpe", "textrecipes")
+  } else if (x$engine == "udpipe") {
+    c("udpipe", "textrecipes")
+  } else {
+    "textrecipes"
+  }
+}
+
+#' @rdname tunable.step
+#' @export
+tunable.step_tokenize <- function(x, ...) {
+  tibble::tibble(
+    name = c("token"),
+    call_info = list(
+      list(pkg = "dials", fun = "token")
+    ),
+    source = "recipe",
+    component = "step_tokenize",
+    component_id = x$id
+  )
 }