+++ tests/testthat/test-hashing.R
@@ -3,13 +3,14 @@ context("test-hashing")
 library(textrecipes)
 library(recipes)
 
-test_data <- tibble(text = c("I would not eat them here or there.",
-                             "I would not eat them anywhere.",
-                             "I would not eat green eggs and ham.",
-                             "I do not like them, Sam-I-am.")
-)
+test_data <- tibble(text = c(
+  "I would not eat them here or there.",
+  "I would not eat them anywhere.",
+  "I would not eat green eggs and ham.",
+  "I do not like them, Sam-I-am."
+))
 
-rec <- recipe(~ ., data = test_data)
+rec <- recipe(~., data = test_data)
 
 test_that("hashing gives double outputs", {
   skip_if_not_installed("text2vec")
@@ -21,13 +22,13 @@ test_that("hashing gives double outputs", {
     prep()
 
   expect_true(
-    juice(obj) %>%
+    bake(obj, new_data = NULL) %>%
       select(contains("hash")) %>%
       lapply(is.double) %>%
       unlist() %>%
       all()
-    )
-  
+  )
+
   expect_equal(dim(tidy(rec, 2)), c(1, 4))
   expect_equal(dim(tidy(obj, 2)), c(1, 4))
 })
@@ -40,7 +41,7 @@ test_that("hashing output width changes accordingly with num_terms", {
     prep()
 
   expect_equal(
-    juice(rec) %>%
+    bake(rec, new_data = NULL) %>%
       select(contains("hash")) %>%
       ncol(),
     256
@@ -49,19 +50,19 @@ test_that("hashing output width changes accordingly with num_terms", {
 
 test_that("hashing output width changes accordingly with num_terms", {
   skip_if_not_installed("text2vec")
-  
-  signed <- recipe(~ ., data = test_data) %>%
+
+  signed <- recipe(~., data = test_data) %>%
     step_tokenize(all_predictors()) %>%
     step_texthash(all_predictors(), num_terms = 2) %>%
     prep() %>%
-    juice()
-  
-  unsigned <- recipe(~ ., data = test_data) %>%
+    bake(new_data = NULL)
+
+  unsigned <- recipe(~., data = test_data) %>%
     step_tokenize(all_predictors()) %>%
     step_texthash(all_predictors(), num_terms = 2, signed = FALSE) %>%
     prep() %>%
-    juice()
-  
+    bake(new_data = NULL)
+
   all(unsigned$text_hash1 == signed$text_hash1)
   all(unsigned$text_hash2 == signed$text_hash2)
   expect_false(all(unsigned$text_hash1 == signed$text_hash1))
@@ -78,4 +79,3 @@ test_that("printing", {
   expect_output(print(rec))
   expect_output(prep(rec, verbose = TRUE))
 })
-